Fix test data race
